compose-sample: Extract the shared channel list factory helper
